+++ src/test/normalizeOptions.test.js
@@ -1,16 +1,16 @@
-const {
-  normalizeOptions,
-  handleBreakpointConfigs,
-  LENGTH_REGEX,
-} = require('../normalizeOptions');
+const { normalizeOptions, LENGTH_REGEX } = require('../normalizeOptions');
 
 /**
- * Nomalize, collect and merge breakpoint configs.
+ * Nomalize config.
  */
 describe('Normalize option value types', () => {
   test('Omits invalid option values', () => {
     expect(normalizeOptions({
-      columns: 'none', gap: '10vw', edge: 2, siteMax: '90',
+      columns: 'none',
+      gap: '10vw',
+      edge: 2,
+      max: '90',
+      base: 'px',
     }))
       .toEqual({});
   });
@@ -20,128 +20,6 @@ describe('Normalize option value types', () => {
       columns: '12',
     }))
       .toEqual({ columns: 12 });
-  });
-
-  test('Collects options with one included breakpoint config', () => {
-    expect(normalizeOptions({
-      columns: 12,
-      edge: '1rem',
-      gap: '0.625rem',
-      siteMax: '90rem',
-      breakpoints: {
-        '48rem': {
-          gap: '1rem',
-          edge: '1.25rem',
-        },
-      },
-    }))
-      .toEqual({
-        columns: 12,
-        edge: '1rem',
-        gap: '0.625rem',
-        siteMax: '90rem',
-        breakpoints: {
-          '48rem': {
-            gap: '1rem',
-            edge: '1.25rem',
-          },
-        },
-      });
-  });
-
-  test('Collects and merges options with multiple breakpoint configs', () => {
-    expect(normalizeOptions({
-      columns: 12,
-      edge: '1rem',
-      gap: '0.625rem',
-      siteMax: '90rem',
-      breakpoints: {
-        '48rem': {
-          gap: '1rem',
-        },
-        '64rem': {
-          edge: '1.25rem',
-        },
-      },
-    }))
-      .toEqual({
-        columns: 12,
-        edge: '1rem',
-        gap: '0.625rem',
-        siteMax: '90rem',
-        breakpoints: {
-          '48rem': {
-            gap: '1rem',
-          },
-          '64rem': {
-            gap: '1rem',
-            edge: '1.25rem',
-          },
-        },
-      });
-  });
-});
-
-describe('Nomalize, collect and merge breakpoint configs', () => {
-  test('Converts breakpoint value with no units to `px`', () => {
-    expect(handleBreakpointConfigs({
-      768: {
-        gap: '1rem',
-        edge: '1.25rem',
-      },
-    }, {}))
-      .toEqual({
-        breakpoints: {
-          '768px': {
-            gap: '1rem',
-            edge: '1.25rem',
-          },
-        },
-      });
-  });
-
-  test('Collects, merges, and sorts breakpoint values declared out-of-order', () => {
-    expect(handleBreakpointConfigs({
-      1024: {
-        gap: '1.25rem',
-      },
-      '768px': {
-        edge: '1.25rem',
-      },
-    }, {}))
-      .toEqual({
-        breakpoints: {
-          '768px': {
-            edge: '1.25rem',
-          },
-          '1024px': {
-            gap: '1.25rem',
-            edge: '1.25rem',
-          },
-        },
-      });
-  });
-
-  test('Collects and merges breakpoint values with `0` overrrides', () => {
-    expect(handleBreakpointConfigs({
-      '48rem': {
-        gap: '1rem',
-      },
-      '64rem': {
-        edge: 0,
-      },
-    }, {}))
-      .toEqual({
-        breakpoints: {
-          '48rem': {
-            gap: '1rem',
-          },
-          '64rem': {
-            gap: '1rem',
-            edge: 0,
-          },
-        },
-      });
   });
 });