+++ test/sharedConfigs.js
@@ -4,49 +4,49 @@
 module.exports = {
   allValues: {
     columns: 16,
-    siteMax: '75rem',
+    max: '75rem',
     edge: '32px',
     gap: '0.625rem', // 0.5859rem
   },
   edgeGap: {
     columns: 12,
-    siteMax: undefined,
+    max: undefined,
     edge: '1rem',
     gap: '10px', // 9.1667px
   },
-  edgeSiteMax: {
+  edgeMax: {
     columns: 16,
-    siteMax: '1024px',
+    max: '1024px',
     edge: '1.25rem',
     gap: undefined,
   },
-  gapSiteMax: {
+  gapMax: {
     columns: 16,
-    siteMax: '60rem',
+    max: '60rem',
     edge: undefined,
     gap: '15px', // 14.0625px
   },
   edgeOnly: {
     columns: 12,
-    siteMax: undefined,
+    max: undefined,
     edge: '20px',
     gap: undefined,
   },
   gapOnly: {
     columns: 12,
-    siteMax: undefined,
+    max: undefined,
     edge: undefined,
     gap: '0.9375rem', // 0.8594rem
   },
-  siteMaxOnly: {
+  maxOnly: {
     columns: 16,
-    siteMax: '1200px',
+    max: 'var(--tmax)',
     edge: undefined,
     gap: undefined,
   },
   columnsOnly: {
     columns: 12,
-    siteMax: undefined,
+    max: undefined,
     edge: undefined,
     gap: undefined,
   },
@@ -54,25 +54,12 @@ module.exports = {
     columns: 'var(--columns)',
     gap: 'var(--gap)',
     edge: 'var(--edge)',
-    siteMax: '90rem',
+    max: '90rem',
   },
   typical: {
     columns: 12,
     gap: '1.25rem',
     edge: '0.625rem',
-    siteMax: '90rem',
-  },
-  typicalWithBreakpoints: {
-    columns: 12,
-    gap: '1.25rem',
-    edge: '0.625rem',
-    breakpoints: {
-      '768px': {
-        gap: '0.625rem',
-      },
-      '1024px': {
-        siteMax: '90rem',
-      },
-    },
+    max: '90rem',
   },
 };