+++ test/tidy-var.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 const run = require('.');
-const { typical, typicalWithBreakpoints } = require('./sharedConfigs');
+const { typical } = require('./sharedConfigs');
 const { VAR_FUNCTION_REGEX } = require('../tidy-var');
 
 /**
@@ -53,35 +53,6 @@ describe('The `tidy-var()` function is replaced with the expected option value',
   );
 
   test(
-    'Replaces `tidy-var` when used as a `tidy-*` property value',
-    () => run(
-      'div { tidy-span: tidy-var(columns); }',
-      // calc((((100vw - 0.625rem * 2) / 12 - 1.1458rem) * 12) + 1.25rem * 11)
-      'div { width: calc(100vw - 1.2496rem); }',
-      typicalWithBreakpoints,
-    ),
-  );
-
-  test(
-    'Replaces `tidy-var` when used as a `tidy-*` function value',
-    () => run(
-      'div { width: tidy-span(tidy-var(columns)); }',
-      // calc((((100vw - 0.625rem * 2) / 12 - 1.1458rem) * 12) + 1.25rem * 11)
-      'div { width: calc(100vw - 1.2496rem); }',
-      typicalWithBreakpoints,
-    ),
-  );
-
-  test(
-    'Replaces the correct value from within a matched breakpoint',
-    () => run(
-      'div { margin-left: tidy-var(gap); } @media (min-width: 900px) { div { margin-left: tidy-var(gap); } }',
-      'div { margin-left: 1.25rem; } @media (min-width: 900px) { div { margin-left: 0.625rem; } }',
-      typicalWithBreakpoints,
-    ),
-  );
-
-  test(
     'Maintains `tidy-var` input as a /* comment */',
     () => run(
       'div { margin-left: tidy-var(gap); }',
@@ -114,8 +85,8 @@ describe('Matches tidy-var() functions', () => {
       ['tidy-var(edge)', 'edge'],
     ],
     [
-      'tidy-var(siteMax)',
-      ['tidy-var(siteMax)', 'siteMax'],
+      'tidy-var(max)',
+      ['tidy-var(max)', 'max'],
     ],
     [
       'tidy-var(columns)',