+++ tidy-property.js
@@ -23,15 +23,14 @@ const OFFSET_REGEX = /tidy-offset-(left|right)/;
  * @param {Object} tidy        An instance of the Tidy class.
  */
 function tidyProperty(declaration, tidy) {
-  const { fullWidthRule, columns, columns: { options } } = tidy;
+  const { columns, columns: { options } } = tidy;
 
   // Replace `tidy-span` declaration with a `width` declaration.
   if ('tidy-span' === declaration.prop) {
     /**
-     * fluid: calc() function based on 100vw base.
-     * full:  calc() function based on `siteMax` base.
+     * Get the span `calc()` function.
      */
-    const { fluid, full } = columns.spanCalc(declaration.value);
+    const calcValue = columns.spanCalc(declaration.value);
 
     const columnDecl = [];
 
@@ -48,20 +47,9 @@ function tidyProperty(declaration, tidy) {
       declaration,
       {
         prop: 'width',
-        value: fluid,
+        value: calcValue,
       },
     ));
-
-    // Conditionally insert a `max-width` declaration.
-    if (undefined !== full) {
-      columnDecl.push(cleanClone(
-        declaration,
-        {
-          prop: 'max-width',
-          value: full,
-        },
-      ));
-    }
 
     declaration.replaceWith(columnDecl);
   }
@@ -73,11 +61,11 @@ function tidyProperty(declaration, tidy) {
      * direction:  The side upon which the offset will be applied.
      */
     const [, direction] = declaration.prop.match(OFFSET_REGEX);
+
     /**
-     * fluid: calc() function based on 100vw base.
-     * full:  calc() function based on `siteMax` base.
+     * Get the offset `calc()` function.
      */
-    const { fluid, full } = columns.offsetCalc(declaration.value);
+    const calcValue = columns.offsetCalc(declaration.value);
 
     // Save the original declaration in a comment for debugging.
     if (
@@ -93,24 +81,12 @@ function tidyProperty(declaration, tidy) {
       declaration,
       {
         prop: `margin-${direction}`,
-        value: fluid,
+        value: calcValue,
       },
     );
 
     // Replace the declaration with the cloned fluid declaration.
     declaration.replaceWith(fluidDecl);
-
-    // Conditionally add a `margin-[left|right]` declaration to the media query.
-    if (undefined !== full && 'rule' === fullWidthRule.type) {
-      // Clone the declaration with `full` declaration overrides.
-      fullWidthRule.append(cleanClone(
-        declaration,
-        {
-          prop: `margin-${direction}`,
-          value: full,
-        },
-      ));
-    }
   }
 }